Ччинит главную ошибку(падал весь проект)
