+++ src/services/process-error-handle.ts
@@ -1,8 +1,19 @@
+import { createAsyncThunk } from '@reduxjs/toolkit';
 import { store } from '../store';
-import { clearErrorAction } from '../store/api-actions';
 import { setError } from '../store/main-process/main-process';
+
+const TIMEOUT_SHOW_ERROR = 2000;
+
+export const clearErrorAction = createAsyncThunk(
+  'data/clearError',
+  () => {
+    setTimeout(
+      () => store.dispatch(setError(null)),
+      TIMEOUT_SHOW_ERROR,
+    );
+  },
+);
 
 export const processErrorHandle = (message: string): void => {
   store.dispatch(setError(message));
-  store.dispatch(clearErrorAction());
 };
